made models more db-friendly and rails-vernacular

+++ lib/usda_nutrient_database/import/downloader.rb
@@ -6,50 +6,74 @@ module UsdaNutrientDatabase
 
       attr_reader :directory, :version
 
-      def initialize(directory = 'tmp/usda', version = 'sr25')
+      class_attribute :usda_hostname; self.usda_hostname = 'http://www.ars.usda.gov'
+
+      def initialize(directory, version)
         @directory = directory
         @version = version
       end
 
       def cleanup
-        UsdaNutrientDatabase.log 'Cleaning up data'
-        FileUtils.rm_rf(directory)
+        unless UsdaNutrientDatabase.configuration.keep_data?
+          UsdaNutrientDatabase.log "Cleaning up data in #{directory}"
+          FileUtils.rm_rf(directory)
+        end
       end
 
       def download_and_unzip
+        return true if File.exists?(files_path('WEIGHT.txt'))
         download and unzip
       end
 
-      def path
-        [
+      def zipfile_basename
+        "#{version}asc.zip"
+      end
+
+      def url_path
+        # https://www.ars.usda.gov/SP2UserFiles/Place/12354500/Data/SR27/dnload/sr27asc.zip
+        File.join(
           'SP2UserFiles', 'Place', '12354500', 'Data', version.upcase,
-          'dnload', "#{version}.zip"
-        ].join('/')
+          'dnload', zipfile_basename)
+      end
+
+      def url
+        File.join(usda_hostname, url_path)
+      end
+
+      def zipfile_path
+        File.join(directory, zipfile_basename)
+      end
+
+      def files_path(filename=nil)
+        File.join(*[directory, version, filename].compact)
       end
 
       def download
-        UsdaNutrientDatabase.log "Downloading USDA data version: #{version}"
-        unless File.directory?(directory)
-          FileUtils.mkdir_p("#{directory}/#{version}")
-        end
-        File.open("#{directory}/#{version}.zip", 'w+b') do |file|
-          file.write connection.get(path).body
+        if File.exists?(zipfile_path) && false
+          UsdaNutrientDatabase.log "#{zipfile_path} exists, skipping"
+        else
+          UsdaNutrientDatabase.log "Downloading USDA data version #{version} from #{url} into #{zipfile_path}"
+          FileUtils.mkdir_p(File.dirname(zipfile_path))
+          File.open(zipfile_path, 'w+b') do |file|
+            file.write connection.get(url_path).body
+          end
         end
       end
 
       def unzip
-        UsdaNutrientDatabase.log 'Unzipping data'
-        Zip::File.open("#{directory}/#{version}.zip") do |zipfile|
+        UsdaNutrientDatabase.log "Unzipping data from #{zipfile_path} into #{files_path}"
+        FileUtils.mkdir_p(files_path)
+        Zip::File.open(zipfile_path) do |zipfile|
           zipfile.each do |file|
-            unless File.exist?("#{directory}/#{version}/#{file.name}")
-              zipfile.extract(file, "#{directory}/#{version}/#{file.name}")
+            unless File.exist?(files_path(file.name))
+              zipfile.extract(file, files_path(file.name))
             end
           end
         end
       end
 
       def connection
-        @connection ||= Faraday.new(url: 'http://www.ars.usda.gov')
+        @connection ||= Faraday.new(url: usda_hostname)
       end
     end
   end
